# docs/openapi.yml
openapi: 3.0.1
info:
  title: Opdex Platform API
  version: 1.0.0
  contact:
    name: Opdex Contributors
    url: https://github.com/Opdex/opdex-v1-api
  description: The Opdex Platform Web API is an interface for querying information from the Opdex contract indexer and quoting Opdex smart contract transactions. The Opdex contract indexer is a job that continually runs and stores relevant broadcast transaction data, in a way that represents the current known state of the Opdex protocol at the present time.
  license:
    name: MIT
    url: https://github.com/Opdex/opdex-v1-api/blob/main/LICENSE
servers:
  - url: https://test-api.opdex.com
...
components:
  securitySchemes:
    opdexAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authorize with a bearer token obtained from SignalR
  responses:
...
    TooManyRequests:
      description: Too many requests
      headers:
        Retry-After:
          schema:
            type: integer
            format: int32
            minimum: 0
            example: 30
          description: Indicates how many seconds to wait before making a follow-up request
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/problemDetails"
          example:
            type: https://httpstatuses.com/429
            title: Too Many Requests
            detail: "Quota exceeded. Maximum allowed: 500 per 60s. Please try again in 30 second(s)."
            status: 429
            extensions:
              traceId: 00-00000000000000000000000000000000-0000000000000000-00
    InternalServerError:
      description: Unexpected error occurred
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/problemDetails"
          example:
            type: https://httpstatuses.com/500
            title: Internal Server Error
            status: 500
            extensions:
              traceId: 00-00000000000000000000000000000000-0000000000000000-00
  schemas:
...
    validationProblemDetails:
      allOf:
        - $ref: "#/components/schemas/problemDetails"
        - type: object
          properties:
            errors:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
                  description: Validation error message
                minItems: 1
              description: Contains invalid properties and associated error messages
    address:
      type: string
      description: An address on the Cirrus network
      minLength: 30
      maxLength: 42
      pattern: ^[a-km-zA-HJ-NP-Z1-9]$
...
    stakingStatus:
      type: string
      description: Staking status filter
      enum:
        - Enabled
        - Disabled
    miningStatus:
      description: Mining status filter
      enum:
        - Enabled
        - Disabled
    nominationStatus:
      description: Nomination status filter
      enum:
        - Nominated
        - Excluded
    liquidityPoolOrderByType:
      type: string
      description: Property to sort liquidity pools by
      enum:
        - Liquidity
        - Volume
        - StakingWeight
        - Name
...
    sortDirection:
      type: string
      description: Sort order
      enum:
        - ASC
        - DESC
...
    liquidityPoolsResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/liquidityPoolResponse"
        paging:
          $ref: "#/components/schemas/pagingResponse"
...
    pagingResponse:
      type: object
      description: Paging cursors
      properties:
        next:
          type: string
          description: Cursor referencing the next page
        previous:
          type: string
          description: Cursor referencing the previous page
...
paths:
...
  /liquidity-pools:
    get:
      tags:
        - Liquidity Pools
      summary: Get Liquidity Pools
      description: Retrieves details for Opdex liquidity pools across one or many markets. This is a [paginated endpoint](https://docs.opdex.com/reference/using-the-opdex-platform-api#paginated-endpoints), so a request will retrieve a single page of results. It is recommended to supply filters to your query, to reduce the total number of requests that you may have to make.
      operationId: getLiquidityPools
      parameters:
        - name: keyword
          in: query
          description: Keyword search against liquidity pool addresses and names
          schema:
            type: string
          example: BTC
        - name: markets
          in: query
          description: Addresses of markets to search within
          schema:
            type: array
            items:
              $ref: "#/components/schemas/address"
              description: Address of a market
          example:
            - t7RorA7xQCMVYKPM1ibPE1NSswaLbpqLQb
        - name: liquidityPools
          in: query
          description: Addresses of liquidity pools to fetch
          schema:
            type: array
            items:
              $ref: "#/components/schemas/address"
              description: Address of a liquidity pool
          example:
            - tMdZ2UfwJorAyErDvqNdVU8kmiLaykuE5L
            - tLrMcU1csbN7RxGjBMEnJeLoae3PxmQ9cr
        - name: tokens
          in: query
          description: Addresses of SRC tokens paired in pools
          schema:
            type: array
            items:
              $ref: "#/components/schemas/address"
              description: Address of an SRC token
          example:
            - tGSk2dVENuqAQ2rNXbui37XHuurFCTqadD
            - tF83sdXXt2nTkL7UyEYDVFMK4jTuYMbmR3
        - name: stakingStatus
          in: query
          description: Staking status filter; default is not to filter
          schema:
            $ref: "#/components/schemas/stakingStatus"
          example: Enabled
        - name: nominationStatus
          in: query
          description: Nomination status filter; default is not to filter
          schema:
            $ref: "#/components/schemas/nominationStatus"
          example: Nominated
        - name: miningStatus
          in: query
          description: Mining status filter; default is not to filter
          schema:
            $ref: "#/components/schemas/miningStatus"
          example: Enabled
        - name: orderBy
          in: query
          description: Property by which to sort results
          schema:
            $ref: "#/components/schemas/liquidityPoolOrderByType"
          example: Volume
        - name: direction
          in: query
          description: Order direction of the results
          schema:
            $ref: "#/components/schemas/sortDirection"
          example: DESC
        - name: limit
          in: query
          description: Number of results per page
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 50
          example: 10
        - name: cursor
          in: query
          description: Reference of the requested page, returned by a previous call
          schema:
            type: string
      responses:
        200:
          description: Liqudidity pool results found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/liquidityPoolsResponse"
              example:
                results:
                  - address: tVFhXcS3gVb49MSTsaDFoqkxLrAUiNEN7n
                    name: TBTC-TCRS
                    transactionFee: 0.3
                    tokens:
                      crs:
                        address: CRS
                        name: Cirrus
                        symbol: TCRS
                        decimals: 8
                        sats: 100000000
                        totalSupply: "100000000.00000000"
                        attributes: []
                        summmary:
                          priceUsd: 1.36598319
                          dailyPriceChangePercent: 0.00000000
                          modifiedBlock: 2988890
                      src:
                        address: tGSk2dVENuqAQ2rNXbui37XHuurFCTqadD
                        name: Bitcoin (Wrapped)
                        symbol: TBTC
                        decimals: 8
                        sats: 100000000
                        totalSupply: "21000000.00000000"
                        attributes:
                          - NonProvisional
                        summmary:
                          priceUsd: 42633.71834591
                          dailyPriceChangePercent: -0.04000000
                          modifiedBlock: 2991118
                      lp:
                        address: tVFhXcS3gVb49MSTsaDFoqkxLrAUiNEN7n
                        name: Opdex Liquidity Pool Token
                        symbol: OLPT
                        decimals: 8
                        sats: 100000000
                        totalSupply: "94.39473361"
                        attributes:
                          - Provisional
                        summmary:
                          priceUsd: 482.64677779
                          dailyPriceChangePercent: -0.02000000
                          modifiedBlock: 2991118
                    summary:
                      reserves:
                        crs: "16676.38165376"
                        src: "0.53431082"
                        usd: 45559.31401812
                        dailyUsdChangePercent: -0.02000000
                      rewards:
                        providerDailyUsd: 0.01000000
                        marketDailyUsd: 0.00000000
                        totalDailyUsd: 0.01000000
                      volume:
                        dailyUsd: 4.26496572
                      cost:
                        crsPerSrc: "31211.01244732"
                        srcPerCrs: "0.00003203"
                      staking:
                        token:
                          market: t7RorA7xQCMVYKPM1ibPE1NSswaLbpqLQb
                          liquidityPool: tGS4ad4E1eNXcKxoke3t7v5zQtjMpeRQ3U
                          address: tTTuKbCR2UnsEByXBp1ynBz91J2yz63h1c
                          name: Testnet Opdex Token
                          symbol: TODX
                          decimals: 8
                          sats: 100000000
                          totalSupply: "400000000.00000000"
                          attributes:
                            - NonProvisional
                            - Staking
                          summary:
                            priceUsd: 0.00000000
                            dailyPriceChangePercent: 0.00000000
                            modifiedBlock: 2989456
                        weight: "0.00000000"
                        usd: 0.00000000
                        dailyWeightChangePercent: 0.00000000
                        nominated: true
                      miningPool:
                        address: tRs6rXfHuLhKZhWuWpycLASzAyn4kXo6bT
                        liquidityPool: tVFhXcS3gVb49MSTsaDFoqkxLrAUiNEN7n
                        miningPeriodEndBlock: 0
                        rewardPerBlock: "0.00000000"
                        rewardPerLpt: "0.00000000"
                        tokensMining: "4.14831968"
                        isActive: false
                paging: {}
        400:
          description: The request is not valid
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/validationProblemDetails"
              example:
                type: https://httpstatuses.com/400
                title: One or more validation errors occurred.
                status: 400
                errors:
                  nominationStatus:
                    - The value 'Enabled' is not valid for NominationStatus.
                extensions:
                  traceId: 00-00000000000000000000000000000000-0000000000000000-00
        429:
          $ref: "#/components/responses/TooManyRequests"
        500:
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - Liquidity Pools
      summary: Build Create Liquidity Pool Transaction Quote
      description: Builds a quote for a transaction to create an Opdex liquidity pool. The quote can be broadcast by a Stratis Transaction Handoff Broadcastor. See the [specification](https://github.com/Opdex/STHS) for further details.
      operationId: buildCreateLiquidityPoolQuote
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/quoteCreateLiquidityPoolRequest"
            example:
              token: tGSk2dVENuqAQ2rNXbui37XHuurFCTqadD
              market: t7RorA7xQCMVYKPM1ibPE1NSswaLbpqLQb
      security:
        - opdexAuth: []
      responses:
        200:
          description: Built create liquidity pool quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/transactionQuoteResponse"
              example:
                result: tNiP367ETVYReXenrZXz5XQGVXt1FFHQz1
                gasUsed: 59204
                events:
                  - token: tA88mLAYt3gTp5eqFR6srceAPrTNf4tdFy
                    liquidityPool: tNiP367ETVYReXenrZXz5XQGVXt1FFHQz1
                    eventType: CreateLiquidityPoolEvent
                    contract: t7RorA7xQCMVYKPM1ibPE1NSswaLbpqLQb
                    sortOrder: 0
                request: eyJzZW5kZXIiOiJ0UTlSdWtac0I2YkJzZW5IbkdTbzFxNjlDSnpXR254b2htIiwidG8iOiJ0N1JvckE3eFFDTVZZS1BNMWliUEUxTlNzd2FMYnBxTFFiIiwiYW1vdW50IjoiMCIsIm1ldGhvZCI6IkNyZWF0ZVBvb2wiLCJwYXJhbWV0ZXJzIjpbeyJsYWJlbCI6IlRva2VuIEFkZHJlc3MiLCJ2YWx1ZSI6IjkjdEE4OG1MQVl0M2dUcDVlcUZSNnNyY2VBUHJUTmY0dGRGeSJ9XSwiY2FsbGJhY2siOiJodHRwczovL3Rlc3QtYXBpLm9wZGV4LmNvbS90cmFuc2FjdGlvbnMifQ==
        400:
          description: The request is not valid
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/validationProblemDetails"
              example:
                type: https://httpstatuses.com/400
                title: One or more validation errors occurred.
                status: 400
                errors:
                  market:
                    - Invalid address.
                extensions:
                  traceId: 00-00000000000000000000000000000000-0000000000000000-00
        401:
          $ref: "#/components/responses/Unauthorized"
        429:
          $ref: "#/components/responses/TooManyRequests"
        500:
          $ref: "#/components/responses/InternalServerError"
  /liquidity-pools/{pool}:
    get:
      tags:
        - Liquidity Pools
      summary: Get Liquidity Pool
      description: Retrieve details for an Opdex liquidity pool.
      operationId: getLiquidityPool
      parameters:
        - name: pool
          in: path
          description: Address of a liquidity pool
          required: true
          schema:
            $ref: "#/components/schemas/address"
          example: tMdZ2UfwJorAyErDvqNdVU8kmiLaykuE5L
      responses:
        200:
          description: Liquidity pool found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/liquidityPoolResponse"
              example:
                address: tVFhXcS3gVb49MSTsaDFoqkxLrAUiNEN7n
                name: TBTC-TCRS
                transactionFee: 0.3
                tokens:
                  crs:
                    address: CRS
                    name: Cirrus
                    symbol: TCRS
                    decimals: 8
                    sats: 100000000
                    totalSupply: "100000000.00000000"
                    attributes: []
                    summmary:
                      priceUsd: 1.36598319
                      dailyPriceChangePercent: 0.00000000
                      modifiedBlock: 2988890
                  src:
                    address: tGSk2dVENuqAQ2rNXbui37XHuurFCTqadD
                    name: Bitcoin (Wrapped)
                    symbol: TBTC
                    decimals: 8
                    sats: 100000000
                    totalSupply: "21000000.00000000"
                    attributes:
                      - NonProvisional
                    summmary:
                      priceUsd: 42633.71834591
                      dailyPriceChangePercent: -0.04000000
                      modifiedBlock: 2991118
                  lp:
                    address: tVFhXcS3gVb49MSTsaDFoqkxLrAUiNEN7n
                    name: Opdex Liquidity Pool Token
                    symbol: OLPT
                    decimals: 8
                    sats: 100000000
                    totalSupply: "94.39473361"
                    attributes:
                      - Provisional
                    summmary:
                      priceUsd: 482.64677779
                      dailyPriceChangePercent: -0.02000000
                      modifiedBlock: 2991118
                summary:
                  reserves:
                    crs: "16676.38165376"
                    src: "0.53431082"
                    usd: 45559.31401812
                    dailyUsdChangePercent: -0.02000000
                  rewards:
                    providerDailyUsd: 0.01000000
                    marketDailyUsd: 0.00000000
                    totalDailyUsd: 0.01000000
                  volume:
                    dailyUsd: 4.26496572
                  cost:
                    crsPerSrc: "31211.01244732"
                    srcPerCrs: "0.00003203"
                  staking:
                    token:
                      market: t7RorA7xQCMVYKPM1ibPE1NSswaLbpqLQb
                      liquidityPool: tGS4ad4E1eNXcKxoke3t7v5zQtjMpeRQ3U
                      address: tTTuKbCR2UnsEByXBp1ynBz91J2yz63h1c
                      name: Testnet Opdex Token
                      symbol: TODX
                      decimals: 8
                      sats: 100000000
                      totalSupply: "400000000.00000000"
                      attributes:
                        - NonProvisional
                        - Staking
                      summary:
                        priceUsd: 0.00000000
                        dailyPriceChangePercent: 0.00000000
                        modifiedBlock: 2989456
                    weight: "0.00000000"
                    usd: 0.00000000
                    dailyWeightChangePercent: 0.00000000
                    nominated: true
                  miningPool:
                    address: tRs6rXfHuLhKZhWuWpycLASzAyn4kXo6bT
                    liquidityPool: tVFhXcS3gVb49MSTsaDFoqkxLrAUiNEN7n
                    miningPeriodEndBlock: 0
                    rewardPerBlock: "0.00000000"
                    rewardPerLpt: "0.00000000"
                    tokensMining: "4.14831968"
                    isActive: false
        400:
          description: The request is not valid
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/validationProblemDetails"
              example:
                type: https://httpstatuses.com/400
                title: One or more validation errors occurred.
                status: 400
                errors:
                  pool:
                    - Invalid address.
                extensions:
                  traceId: 00-00000000000000000000000000000000-0000000000000000-00
        404:
          $ref: "#/components/responses/LiquidityPoolNotFound"
        429:
          $ref: "#/components/responses/TooManyRequests"
        500:
          $ref: "#/components/responses/InternalServerError"
...